src/modules/ibc/ics27.ts
import { MsgDelegate, Msg, Coin, LCDClient } from "@terra-money/feather.js";

// Define the MsgExecute structure
export interface MsgExecute {
  sender: string;
  interchainAccount: string;
  msgs: Msg[];
}

export class ICS27 {
  private lcdClient: LCDClient;

  constructor(lcdClient: LCDClient) {
    this.lcdClient = lcdClient;
  }

  async transferTokens(recipientAddress: string, amount: number) {
    // Implementation for performing transfers of ICS20 tokens
    return `Success message ${recipientAddress}, ${amount}, ${this.lcdClient}`;
  }

  // Function to create MsgExecute
  createMsgExecute(
    sender: string,
    interchainAccount: string,
    msgs: Msg[]
  ): MsgExecute {
    return {
      sender,
      interchainAccount,
      msgs,
    };
  }

  // Function to encode MsgExecute
  encodeMsgExecute() {
    // Encoding logic for MsgExecute
  }

  // Function to decode MsgExecute
  decodeMsgExecute() {
    // Decoding logic for MsgExecute
  }

  createInterchainDelegationMessage(
    sender: string,
    interchainAccount: string,
    validatorAddress: string,
    amount: number,
    denom: string
  ): MsgExecute {
    const msgDelegate = new MsgDelegate(
      interchainAccount,
      validatorAddress,
      new Coin(amount.toString(), denom)
    );

    return this.createMsgExecute(sender, interchainAccount, [msgDelegate]);
  }
}
